backend/requirements.txt
# backend/requirements.txt

# ===== FRAMEWORK PRINCIPAL =====
fastapi==0.104.1
uvicorn[standard]==0.24.0

# ===== BASE DE DATOS =====
motor==3.3.2
pymongo==4.6.0

# ===== AUTENTICACIÓN Y SEGURIDAD =====
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# ===== VALIDACIÓN DE DATOS =====
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0

# ===== CONFIGURACIÓN =====
python-decouple==3.8
python-dotenv==1.0.0

# ===== MANEJO DE ARCHIVOS =====
python-magic==0.4.27
Pillow==10.1.0
aiofiles==23.2.1

# ===== LOGGING Y MONITOREO =====
structlog==23.2.0

# ===== UTILIDADES =====
python-dateutil==2.8.2
pytz==2023.3

# ===== DESARROLLO Y TESTING =====
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# ===== DOCUMENTACIÓN =====
# FastAPI incluye automáticamente:
# - swagger-ui-bundle
# - redoc

# ===== OPCIONAL (REDIS para cache en futuras fases) =====
# redis==5.0.1
# aioredis==2.0.1

# ===== PROCESAMIENTO DE EXCEL (para Fase 2) =====
# pandas==2.1.4
# openpyxl==3.1.2
# xlsxwriter==3.1.9

# ===== EMAIL (para notificaciones futuras) =====
# aiosmtplib==3.0.1
# jinja2==3.1.2

# ===== VERSIONES ESPECÍFICAS PARA COMPATIBILIDAD =====
typing-extensions>=4.8.0
anyio>=3.7.1
sniffio>=1.3.0
idna>=3.4
certifi>=2023.7.22
charset-normalizer>=3.3.0
urllib3>=2.0.7

# ===== DEPENDENCIAS DEL SISTEMA =====
# Estas se instalan a nivel de sistema en el Dockerfile:
# - libmagic1 (para python-magic)
# - libjpeg-dev (para Pillow)
# - libpng-dev (para Pillow)
# - libfreetype6-dev (para Pillow)